Day 9/Mentor_AI/requirements.txt
streamlit
langchain
langchain-google-genai
google-generativeai
faiss-cpu
chromadb
python-dotenv
pydantic
typing-extensions